requirements.txt
django==4.2.23
django-autocomplete-light==3.9.7
django-braces==1.16.0
django-crispy-forms==2.3
django-dbbackup==4.2.1
django-debug-toolbar==5.2.0
django-filter==24.3
django-localflavor==4.0
crispy-bootstrap5==2025.4
psycopg2==2.9.10
gunicorn